% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LoadBackgroundErrorRate}
\alias{LoadBackgroundErrorRate}
\title{function LoadBackgroundErrorRate}
\usage{
LoadBackgroundErrorRate(pos_ranges.file, ber.ctrl.file, n.trim)
}
\arguments{
\item{pos_ranges.file, }{char, name of the Rdata file containing the three variables \code{pos_ind}, \code{pos_snp}, \code{pos_ranges} as build by the function \code{PrepareLibrary}. Default NULL, use the position_ranges.rda provides that we used for our analysis.}

\item{ber.ctrl.file, }{char, pathname of the file providing the background error rates obtained from the controls (default NULL use the provided background error rates obtained from our 29 controls). See \code{background_error_rate.txt} data and \code{BuildCtrlErrorRate} function.}

\item{n.trim, }{integer, number of base positions trimmed at the ends of each amplicon (default 8)}
}
\value{
the adapted background error rate
}
\description{
This function will load the background error rates created from the controls using the function BuildCtrlErrorRate
}
\references{
\emph{Analysis of base-position error rate of next-generation sequencing to detect tumor mutations in circulating DNA} N. Pécuchet, Y. Rozenholc, E. Zonta, D. Pietraz, A. Didelot, P. Combe, L. Gibault, J-B. Bachet, V. Taly, E. Fabre, H. Blons, P. Laurent-Puig in \emph{Clinical Chemistry}
}
\author{
N. Pécuchet, P. Laurent-Puig and Y. Rozenholc
}
