% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{UpperPart}
\alias{UpperPart}
\title{function UpperPart}
\usage{
UpperPart(values, q)
}
\arguments{
\item{values, }{numeric, the set of values}

\item{q, }{the upper quantile}
}
\value{
the indices of the highest values, larger than the given quantile
}
\description{
Find the locations inside a set of values higher than an upper quantile of these values
}
